Request number of page views from GA in batches
We were sending a list of all of the content items in an organisation
to Google Analytics and requesting the number of page views for them all.

This is ok for organisations with a smaller number of content items, but
for organisations with content items in the tens of thousands, the
request fails.

To fix this, the request to GA has been wrapped in a batch. The batch
size has been hard-coded to 1 to get the test to pass, but will made to
be configurable in the next commit.

Remove unnecessary assignment in the existing test. Assigning content items
to the organisation does affect the outcome of the test, so it has been
removed.