Make batch size a constant
In a previous commit the batch size was given a default value.
This commit changes that into a constant value. This makes the size easier
to read.

Through trial and error, I’ve hit upon a batch size of 3000 for one
month. This size could be larger for 7 days page views, and smaller for
one year’s page views.